Creates group leave service
