+++ src/app/core/http-services/group-leave.service.ts
@@ -1,0 +1,27 @@
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { appConfig } from 'src/app/shared/helpers/config';
+import { ActionResponse, successData } from 'src/app/shared/http-services/action-response';
+
+interface LeaveGroupResponse {
+  changed: boolean
+}
+
+@Injectable({
+  providedIn: 'root'
+})
+export class GroupLeaveService {
+
+  constructor(private http: HttpClient) {}
+
+  leave(groupId: string): Observable<boolean> {
+    return this.http.delete<ActionResponse<LeaveGroupResponse>>(`${appConfig().apiUrl}/current-user/group-memberships/${groupId}`)
+      .pipe(
+        map(successData),
+        map((data: LeaveGroupResponse) => data.changed)
+      );
+  }
+
+}
